Add async logging support
