+++ Logger/logger_internal.hpp
@@ -6,6 +6,7 @@
 #include "../CPP_class/class_string_class.hpp"
 #include "../Errno/errno.hpp"
 #include "logger.hpp"
+#include "logger_log_async.hpp"
 
 extern ft_logger *g_logger;
 extern t_log_level g_level;
@@ -33,6 +34,5 @@ void ft_log_rotate(s_file_sink *sink);
 void ft_file_sink(const char *message, void *user_data);
 const char *ft_level_to_str(t_log_level level);
 void ft_log_vwrite(t_log_level level, const char *fmt, va_list args);
-void ft_log_enqueue(t_log_level level, const char *fmt, va_list args);
 
 #endif
