+++ Test/test_logger_async.cpp
@@ -1,0 +1,27 @@
+#include "../Logger/logger.hpp"
+#include "../System_utils/test_runner.hpp"
+#include "../Libft/libft.hpp"
+#include <unistd.h>
+
+FT_TEST(test_logger_async, "logger async logging")
+{
+    int pipe_fds[2];
+    int write_fd;
+    ssize_t read_count;
+    char buffer[256];
+
+    FT_ASSERT_EQ(0, pipe(pipe_fds));
+    write_fd = pipe_fds[1];
+    FT_ASSERT_EQ(0, ft_log_add_sink(ft_json_sink, &write_fd));
+    ft_log_enable_async(true);
+    ft_log_info("async check");
+    ft_log_enable_async(false);
+    read_count = read(pipe_fds[0], buffer, sizeof(buffer) - 1);
+    FT_ASSERT(read_count > 0);
+    buffer[read_count] = '\0';
+    FT_ASSERT(ft_strstr(buffer, "async check") != ft_nullptr);
+    ft_log_remove_sink(ft_json_sink, &write_fd);
+    close(pipe_fds[0]);
+    close(pipe_fds[1]);
+    return (1);
+}
